Split client restriction supportability tests
